Testeo_segunda_parte

+++ bonita_smile_v1/Servicios/Clinicas.cs
@@ -25,12 +25,11 @@ namespace bonita_smile_v1.Servicios
         }
 
 
-        public List<PermisosModel> Mostrar_Permisos_socio(int id_rol,List<string>lista)
+        public List<PermisosModel> Mostrar_Permisos_socio(int id_rol,string alias)
         {
             List<PermisosModel> listaPermisos = new List<PermisosModel>();
-            foreach(var id in lista)
-            {
-                query = "select clinica.nombre_sucursal,clinica.id_clinica,usuario.nombre,usuario.apellidos,usuario.id_usuario,usuario.alias from usuario left join permisos on usuario.id_usuario=permisos.id_usuario left join clinica on clinica.id_clinica=permisos.id_clinica inner join rol on rol.id_rol=usuario.id_rol where rol.id_rol=" + id_rol+" and clinica.id_clinica='"+id+"'";
+           
+                query = "(select Distinct '' as nombre_sucursal,'' as id_clinica,usuario.id_usuario,usuario.alias,usuario.nombre,usuario.apellidos,usuario.password,usuario.id_rol,rol.descripcion from usuario inner join permisos on usuario.id_usuario=permisos.id_usuario inner join rol on usuario.id_rol=rol.id_rol where id_clinica='' and usuario.id_rol="+id_rol+") union (select  DISTINCT clinica.nombre_sucursal,clinica.id_clinica,usuario.id_usuario,usuario.alias,usuario.nombre,usuario.apellidos,usuario.password,usuario.id_rol,rol.descripcion from usuario inner join rol on usuario.id_rol=rol.id_rol INNER join permisos on permisos.id_usuario=usuario.id_usuario inner join clinica on clinica.id_clinica=permisos.id_clinica where permisos.id_clinica in (select id_clinica from usuario inner join permisos on usuario.id_usuario = permisos.id_usuario where usuario.alias='"+alias+"') AND usuario.id_rol="+id_rol+")";
 
                 try
                 {
@@ -47,10 +46,10 @@ namespace bonita_smile_v1.Servicios
                        
                         permisosModel.nombre_sucursal = reader[0].ToString();
                         permisosModel.id_clinica = reader[1].ToString();
-                        permisosModel.nombre = reader[2].ToString();
-                        permisosModel.apellidos = reader[3].ToString();
-                        permisosModel.id_usuario = reader[4].ToString();
-                        permisosModel.alias = reader[5].ToString();
+                        permisosModel.id_usuario = reader[2].ToString();
+                        permisosModel.alias = reader[3].ToString();
+                        permisosModel.nombre = reader[4].ToString();
+                        permisosModel.apellidos = reader[5].ToString();
 
 
                         listaPermisos.Add(permisosModel);
@@ -61,7 +60,7 @@ namespace bonita_smile_v1.Servicios
                     MessageBox.Show(ex.ToString());
                 }
                 conexionBD.Close();
-            }
+            
             
             return listaPermisos;
 
@@ -182,7 +181,7 @@ namespace bonita_smile_v1.Servicios
             }
         }
 
-        public bool eliminar_Permiso(string id_usuario)
+        public bool eliminar_Permiso(string id_usuario,string id_clinica)
         {
 
             bool internet = ti.Test();
@@ -206,7 +205,75 @@ namespace bonita_smile_v1.Servicios
                 }
                 else
                 {
-                    query = "DELETE FROM permisos where id_usuario='" + id_usuario + "'";
+                    int no_permisos = 0;
+                    query = "select IF((select count(permisos.id_usuario) from permisos where permisos.id_usuario='" + id_usuario + "')>1,1,0)";
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    reader = cmd.ExecuteReader();
+
+                    while (reader.Read())
+                    {
+                        no_permisos = Int32.Parse(reader[0].ToString());
+                    }
+                    conexionBD.Close();
+                    if (no_permisos==1)
+                    {
+                        query = "DELETE FROM permisos where id_usuario='" + id_usuario + "' and id_clinica='" + id_clinica + "'";
+                        
+                    }
+                    else
+                    {
+                        query = "DELETE FROM permisos where id_usuario='" + id_usuario + "' and id_clinica='" + id_clinica + "';";
+                        query = query + "insert into permisos (id_usuario) VALUES('" + id_usuario + "')";
+                    }
+                   
+
+                    conexionBD.Open();
+
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
+                    Escribir_Archivo ea = new Escribir_Archivo();
+                    ea.escribir(query + ";");
+                }
+                return true;
+            }
+            catch (MySqlException ex)
+            {
+                System.Windows.MessageBox.Show(ex.ToString());
+                conexionBD.Close();
+                return false;
+            }
+        }
+
+        public bool eliminar_Permisos(string id_usuario)
+        {
+
+            bool internet = ti.Test();
+            try
+            {
+
+                MySqlCommand cmd; ;
+                if (online)
+                {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+
+                    query = "DELETE FROM permisos where id_usuario='" + id_usuario+"';";
+                    query = query + "insert into permisos (id_usuario) VALUES('" + id_usuario + "');";
 
                     conexionBD.Open();
                     cmd = new MySqlCommand(query, conexionBD);
@@ -368,7 +435,7 @@ namespace bonita_smile_v1.Servicios
             }
         }
 
-        public bool actualizar_Permisos(string id_usuario, string id_clinica)
+        public bool actualizar_Permisos(string id_usuario, string id_clinica,string id_clinica_anterior)
         {
 
             bool internet = ti.Test();
@@ -394,8 +461,9 @@ namespace bonita_smile_v1.Servicios
                 else
                 {
                     //string auxiliar_identificador = MostrarUsuario_Update(id_usuario);
-                    query = "UPDATE permisos set id_clinica = '" + id_clinica  + "',auxiliar_identificador ='" + id_usuario + "' where id_usuario ='" + id_usuario + "'";
+                    query = "UPDATE permisos set id_clinica = '" + id_clinica  + "' where id_usuario ='" + id_usuario + "' and id_clinica='"+ id_clinica_anterior + "'";
                     Console.WriteLine(query);
+                    System.Windows.MessageBox.Show(query);
                     conexionBD.Open();
                     cmd = new MySqlCommand(query, conexionBD);
                     cmd.ExecuteReader();
